cache: configure using total memory usage
The existing pattern for vttablet/vtgate cache configuration is a
dangerous practice, because it lets the user configure the number of
items that can be stored in the cache, as opposed to the total amount of
memory (approximately) that the cache will consume. This makes tuning
production systems complicated, and will skew more complex cache
implementations that use size-aware eviction policies.

To fix this, we're deprecating the original config settings for cache
tuning, and introducing new ones where the total size of the cache is
defined in BYTES as opposed to ENTRIES.

To maintain backwards compatibility, if the user supplies the legacy
config options with number of ENTRIES, we'll calculate an approximate
total size for the cache based on the average size of a cache entry for
each given cache.

Signed-off-by: Vicent Marti <vmg@strn.cat>